Modularization refactoring
